+++ Game/src/Components/Game.java
@@ -34,7 +34,6 @@ public class Game extends Canvas implements Runnable {
 		
 		handler.addObject(new Player(WIDTH/2-32, HEIGHT/2-32, ID.Player, handler));
 		handler.addObject(new BasicEnemy(r.nextInt(WIDTH), r.nextInt(HEIGHT), ID.BasicEnemy, handler));
-	
 	}
 
 	
@@ -111,7 +110,7 @@ public class Game extends Canvas implements Runnable {
 		
 	}
 	
-	public static int clamp(int var, int min, int max) {
+	public static float clamp(float var, float min, float max) {
 		if(var >= max)
 			return var = max;
 		else if(var <= min)
